requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
pdfplumber
python-docx
openai
jinja2
aiofiles
